Update README.md
Our project is about the debate between infcluencers, Tana and Brooke and their "scandal" over their hsared Podcast Cancelled. Our concepts were 1=pro-Tana and 2=pro-Brooke over a tiktok debriefing their drama. Maliika and I expected the comments to be pro-Tana as she has the larger audience, is the reason for Brooke's fame, and is a less controversial character. However, we were surprised because support was overhwelmingly for Brooke. Our results challenge the notions of the trend of cancel culture to lack nuance, as even fans who did not like Brooke took her side in this debate.